+++ examples/inspector_query.rs
@@ -1,19 +1,24 @@
 use bevy::prelude::*;
-use bevy_inspector_egui::{widgets::InspectorQuery, Inspectable, InspectorPlugin};
+use bevy_inspector_egui::widgets::{InspectorQuery, InspectorQuerySingle};
+use bevy_inspector_egui::{Inspectable, InspectorPlugin};
 
 #[derive(Inspectable, Default)]
 struct Data {
     query: InspectorQuery<Entity, With<Transform>>,
     has_material: InspectorQuery<&'static mut Handle<StandardMaterial>>,
+    player: InspectorQuerySingle<Entity, With<Player>>,
 }
 
 fn main() {
     App::build()
         .add_plugins(DefaultPlugins)
         .add_plugin(InspectorPlugin::<Data>::new())
+        .add_plugin(InspectorPlugin::<InspectorQuerySingle<Entity, With<Player>>>::new())
         .add_startup_system(setup.system())
         .run();
 }
+
+struct Player;
 
 fn setup(
     mut commands: Commands,
@@ -25,12 +30,14 @@ fn setup(
         material: materials.add(Color::rgb(0.3, 0.5, 0.3).into()),
         ..Default::default()
     });
-    commands.spawn_bundle(PbrBundle {
-        mesh: meshes.add(Mesh::from(shape::Cube { size: 1.0 })),
-        material: materials.add(Color::rgb(0.8, 0.7, 0.6).into()),
-        transform: Transform::from_xyz(0.0, 0.5, 0.0),
-        ..Default::default()
-    });
+    commands
+        .spawn_bundle(PbrBundle {
+            mesh: meshes.add(Mesh::from(shape::Cube { size: 1.0 })),
+            material: materials.add(Color::rgb(0.8, 0.7, 0.6).into()),
+            transform: Transform::from_xyz(0.0, 0.5, 0.0),
+            ..Default::default()
+        })
+        .insert(Player);
     commands.spawn_bundle(LightBundle {
         transform: Transform::from_xyz(4.0, 8.0, 4.0),
         ..Default::default()