+++ src/widgets/mod.rs
@@ -6,5 +6,6 @@ pub use button::InspectableButton;
 pub use new_window::InNewWindow;
 
 pub use crate::reflect::ReflectedUI;
-pub use crate::world_inspector::InspectorQuery;
+pub use crate::world_inspector::impls::InspectorQuery;
+pub use crate::world_inspector::impls::InspectorQuerySingle;
 pub use resource_inspector::ResourceInspector;